Update carbon calc queries to return full information
